django-api-rest/myapp/api/v1/viewsets.py
from rest_framework import viewsets
from rest_framework.permissions import IsAuthenticated
from myapp.models import Client
from .serializers import ClientSerializer


class ClientViewSet(viewsets.ModelViewSet):
    queryset = Client.objects.all()
    serializer_class = ClientSerializer
    permission_classes = [IsAuthenticated]
